Feature/db migration

+++ app/api/collection/models.py
@@ -2,7 +2,7 @@
 Models for collections
 '''
 
-from sqlalchemy import Column, Integer, String, DateTime
+from sqlalchemy import Column, String, DateTime, BigInteger
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 from app.db.base import Base
@@ -14,7 +14,7 @@ class Collection(Base):
     '''
     __tablename__ = "collections"
 
-    game_id = Column(Integer, primary_key=True, index=True)
+    game_id = Column(BigInteger, primary_key=True, index=True)
     game_name = Column(String)
     game_status = Column(String)
     expires_at = Column(DateTime(timezone=True))
